requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Django==4.2
django-autoslug==1.9.9
django-debug-toolbar==4.2.0
django-filter==23.5
django-js-asset==2.2.0
django-lifecycle==1.1.2
django-mptt==0.16.0
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
exceptiongroup==1.2.0
Faker==23.0.0
flake8==7.0.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1
uWSGI==2.0.23
elasticsearch==8.11.0
django-elasticsearch-dsl==8.0
elasticsearch-dsl==8.11.0